make singer_sdk a poetry plugin
