+++ cookiecutter/tap-template/{{cookiecutter.tap_id}}/{{cookiecutter.library_name}}/tests/{{ 'test' }}_core.py
@@ -4,8 +4,7 @@ import datetime
 
 from singer_sdk.testing import (
     TapTestRunner,
-    get_test_class,
-    pytest_generate_tests  # pytest hook function, required for standard tests
+    get_test_class  # pytest hook function, required for standard tests
 )
 from singer_sdk.testing.suites import (
     tap_stream_attribute_tests,
