+++ cookiecutter/target-template/{{cookiecutter.target_id}}/{{cookiecutter.library_name}}/tests/conftest.py
@@ -1,0 +1,3 @@
+"""Test Configuration."""
+
+pytest_plugins = ("singer_sdk.testing.pytest_plugin",)
