+++ pyproject.toml
@@ -192,5 +192,4 @@ requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-# Sample CLI declaration:
-# sdk-tap-countries-sample = 'singer_sdk.samples.sample_tap_countries.countries_tap:SampleTapCountries.cli'
+pytest11 = { callable = "singer_sdk:testing.pytest_plugin", extras = ["testing"] }
