+++ singer_sdk/testing/__init__.py
@@ -1,6 +1,6 @@
 """Tools and standard tests for Tap/Target implementations."""
 
-from .factory import get_test_class, pytest_generate_tests
+from .factory import get_test_class
 from .legacy import (
     _get_tap_catalog,
     _select_all,
@@ -15,7 +15,6 @@ from .runners import SingerTestRunner, TapTestRunner, TargetTestRunner
 
 __all__ = [
     "get_test_class",
-    "pytest_generate_tests",
     "_get_tap_catalog",
     "_select_all",
     "get_standard_tap_tests",
