+++ singer_sdk/testing/factory.py
@@ -9,36 +9,6 @@ import pytest
 from .runners import TapTestRunner, TargetTestRunner
 
 
-def pytest_generate_tests(metafunc: pytest.Metafunc) -> None:
-    """Pytest Hook, responsible for parameterizing tests.
-
-    Called once per each test function, this hook will check if the function name is
-    registered in the parent classes 'params' dict, and if so will parameterize
-    the given test function with the values therein.
-
-    Args:
-        metafunc: Pytest MetaFunc instance, representing a test function or method.
-    """
-    if metafunc.cls and hasattr(metafunc.cls, "params"):
-        funcarglist = metafunc.cls.params.get(metafunc.definition.name)
-        funcargids = (
-            metafunc.cls.param_ids.get(metafunc.definition.name)
-            if hasattr(metafunc.cls, "param_ids")
-            else None
-        )
-        if funcarglist:
-            arg_names = list(funcarglist[0].keys())
-            parameters = [
-                pytest.param(*tuple(funcargs[name] for name in arg_names))
-                for funcargs in funcarglist
-            ]
-            metafunc.parametrize(
-                ",".join(arg_names),
-                parameters,
-                ids=funcargids,
-            )
-
-
 def get_test_class(
     test_runner: TapTestRunner | TargetTestRunner, test_suites: list
 ) -> object:
