+++ tests/external/test_tap_google_analytics.py
@@ -4,7 +4,7 @@ import warnings
 
 from samples.sample_tap_google_analytics.ga_tap import SampleTapGoogleAnalytics
 from singer_sdk.exceptions import ConfigValidationError
-from singer_sdk.testing import TapTestRunner, get_test_class, pytest_generate_tests
+from singer_sdk.testing import TapTestRunner, get_test_class
 from singer_sdk.testing.suites import (
     tap_stream_attribute_tests,
     tap_stream_tests,
